Revert "Se han añadido métodos a los servicios"
