+++ webapp/src/components/Service/LocationService.js
@@ -1,3 +1,4 @@
+import React, { useState, useEffect } from "react";
 import {db} from '../../api/firebase'
